Bug 1213146 - IonMonkey: MIPS: Modify the last two args of Assembler::bind to generic type. r=arai
---
 js/src/jit/mips-shared/Assembler-mips-shared.h | 2 +-
 js/src/jit/mips32/Assembler-mips32.cpp         | 2 +-
 js/src/jit/mips32/Assembler-mips32.h           | 2 +-
 js/src/jit/mips64/Assembler-mips64.cpp         | 2 +-
 js/src/jit/mips64/Assembler-mips64.h           | 2 +-
 5 files changed, 5 insertions(+), 5 deletions(-)

+++ js/src/jit/mips-shared/Assembler-mips-shared.h
@@ -1030,7 +1030,7 @@ class AssemblerMIPSShared : public AssemblerShared
 
     // label operations
     void bind(Label* label, BufferOffset boff = BufferOffset());
-    virtual void bind(InstImm* inst, uint32_t branch, uint32_t target) = 0;
+    virtual void bind(InstImm* inst, uintptr_t branch, uintptr_t target) = 0;
     virtual void Bind(uint8_t* rawCode, AbsoluteLabel* label, const void* address) = 0;
     uint32_t currentOffset() {
         return nextOffset().getOffset();
